Removed outside dependency
